Fixed double addition of row heights while rendering week rows
This addition is already handled in add_week_row method so performing it in add_weeks as well causes table heights to be twice as high as they should be. The effect is probably not too noticeable until multiple event rows are added forcing the row height to increase more than it should.