Refactored calendar class to avoid generating inline colour styles
This is quite a significant change but it allows a number of benefits. I'm not sold on the concept of defining colours within a model - that's really a view concern instead. More than that, it's a styling concern so should really be outside of the markup to keep things semantic.

So I've removed the inline background-color attributes for all day events and color attributes for short events and added some CSS rules to set styling. Because we don't have inline styles any more the CSS works properly but we still need to handle highlighting on hover.

For short events this is simple as a :hover rule on the anchor tag. For all day events I've changed the j/s to add/remove an additional class ec-hover so we can style accordingly. That also means we need only half of the previous j/s.

As a result of this change the color model attribute is no longer needed/required - instead we use the model class name to derive a corresponding CSS class and style that as required. It's still worth having a bit more flexibility in styling so I'll look at adding something for that next.